+++ src/main/dll/dll_000B_dll0b.c
@@ -1449,7 +1449,7 @@ void dll_0B_func0F(int p1, u8 p2, u8 p3, int p4, int p5)
     memset(&gModgfxSpawnContext, 0, sizeof(gModgfxSpawnContext));
     gModgfxSpawnContext.modeByte = p2;
     gModgfxSpawnContext.attachedSource = (void*)p1;
-    gModgfxSpawnContext.sourceModeCopy = (u8)p2;
+    gModgfxSpawnContext.sourceModeCopy = p2;
     fz = lbl_803DF430;
     gModgfxSpawnContext.posX = fz;
     gModgfxSpawnContext.posY = fz;
