+++ src/main/dll/dll_0031_minimap.c
@@ -732,10 +732,10 @@ void Minimap_release(void)
     null = NULL;
     while ((u32)i < 2)
     {
-        if (slots[(u8)i] != NULL)
+        if (slots[i] != NULL)
         {
-            Obj_FreeObject(slots[(u8)i]);
-            slots[(u8)i] = null;
+            Obj_FreeObject(slots[i]);
+            slots[i] = null;
         }
         i++;
     }
